Cambios echos, se agregó la ventana de confirmacion para eliminar y la clase LogUsuario

+++ src/madreteresacrud/TipoGastoABM.java
@@ -8,6 +8,7 @@ import java.beans.Beans;
 import java.util.ArrayList;
 import java.util.List;
 import javax.persistence.RollbackException;
+import javax.swing.JOptionPane;
 import javax.swing.JPanel;
 
 /**
@@ -262,19 +263,23 @@ public class TipoGastoABM extends JPanel {
     }//GEN-LAST:event_refreshButtonActionPerformed
     
     private void deleteButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_deleteButtonActionPerformed
-        int[] selected = masterTable.getSelectedRows();
-        List<madreteresacrud.TipoGasto> toRemove = new ArrayList<madreteresacrud.TipoGasto>(selected.length);
-        for (int idx = 0; idx < selected.length; idx++) {
-            madreteresacrud.TipoGasto t = list.get(masterTable.convertRowIndexToModel(selected[idx]));
-            toRemove.add(t);
-            entityManager.remove(t);
-        }
-        try {
-                entityManager.getTransaction().commit();
-                entityManager.getTransaction().begin();
-            } catch (Exception e) {
+     int reply = JOptionPane.showConfirmDialog(null, "¿Está seguro de eliminar el registro?", "Eliminacion de Registro", JOptionPane.YES_NO_OPTION);
+        if (reply == JOptionPane.YES_OPTION) {
+
+            int[] selected = masterTable.getSelectedRows();
+            List<madreteresacrud.TipoGasto> toRemove = new ArrayList<madreteresacrud.TipoGasto>(selected.length);
+            for (int idx = 0; idx < selected.length; idx++) {
+                madreteresacrud.TipoGasto t = list.get(masterTable.convertRowIndexToModel(selected[idx]));
+                toRemove.add(t);
+                entityManager.remove(t);
             }
-        list.removeAll(toRemove);
+            try {
+                    entityManager.getTransaction().commit();
+                    entityManager.getTransaction().begin();
+                } catch (Exception e) {
+                }
+            list.removeAll(toRemove);
+        }
     }//GEN-LAST:event_deleteButtonActionPerformed
     
     private void newButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_newButtonActionPerformed
